Refactor to isolate testing code from Docker container.
