+++ modules/errored_datums_reader/tests/test_reader.py
@@ -1,29 +1,27 @@
 #!/usr/bin/env python3
-import os
 from collections import defaultdict
 import unittest
 
 from pachyderm_sdk import Client
 
-from errored_datums_reader import reader_main
+from errored_datums_reader import app
 from errored_datums_reader.reader import read_error_files
+from errored_datums_reader.tests import db_connector_iam
+from errored_datums_reader.tests.db_connector_iam import read_environment
 
 
 class ReaderTest(unittest.TestCase):
 
-    @staticmethod
+    def setUp(self):
+        self.client = Client.from_config()
+        self.db = db_connector_iam.connect(read_environment())
+
     @unittest.skip('Requires Pachyderm client configuration.')
-    def test_read_error_files() -> None:
-        client = Client.from_config()
-        files_by_pipeline: defaultdict[str, list[str]] = read_error_files(client)
+    def test_read_error_files(self) -> None:
+        files_by_pipeline: defaultdict[str, list[str]] = read_error_files(self.client)
         for pipeline_name in files_by_pipeline.keys():
             print(f'{pipeline_name} error_count: {len(files_by_pipeline[pipeline_name])}')
 
-    @staticmethod
     @unittest.skip('Integration test skipped due to long process time.')
-    def test_main() -> None:
-        os.environ['DB_CONFIG_SOURCE'] = 'iam'  # Use IAM authentication after 'gcloud auth login'.
-        os.environ['CLIENT_SOURCE'] = 'config'  # Use local ~/.pachyderm/config.json.
-        os.environ['LOG_LEVEL'] = 'DEBUG'
-        os.environ['AUTHORIZATION_TOKEN'] = ''  # Not needed on local machine after 'pachctl auth login'.
-        reader_main.main()
+    def test_app(self) -> None:
+        app.run(self.client, self.db)
